+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 1.2.0
 
 * Add minitest adapter.
+* Fix bug with missing global time execution when tests took less than second.
 
 https://github.com/ArturT/knapsack/compare/v1.1.1...v1.2.0